# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SHIELD API
  version: 0.2.0
  description: |
    AI Security Gateway — analyze prompts for security risks before they reach LLMs.

    Report language resolution order:
    1. `language` in request body
    2. `Accept-Language` header
    3. Auto-detection from prompt text
    4. Default `en`

servers:
  - url: http://localhost:3001
    description: Local development

paths:
  /health:
    get:
      summary: Health check
      tags: [System]
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /ready:
    get:
      summary: Readiness probe
      tags: [System]
      responses:
        '200':
          description: Service is ready

  /live:
    get:
      summary: Liveness probe
      tags: [System]
      responses:
        '200':
          description: Service is live

  /api/v1/openapi.yaml:
    get:
      summary: OpenAPI specification
      tags: [System]
      responses:
        '200':
          description: YAML OpenAPI document
          content:
            text/yaml:
              schema:
                type: string

  /api/v1/analyze:
    post:
      summary: Analyze a prompt for security risks
      tags: [Analysis]
      parameters:
        - name: Accept-Language
          in: header
          required: false
          schema:
            type: string
          example: uk, en;q=0.9
          description: Preferred report language (en or uk)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeRequest'
      responses:
        '200':
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResult'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/analyze/batch:
    post:
      summary: Analyze multiple prompts in one request
      tags: [Analysis]
      description: |
        Runs up to `BATCH_MAX_ITEMS` (default 50) prompts with bounded concurrency.
        Each item is analyzed independently; partial failures are returned per item.
        Counts as a single request for rate limiting.
      parameters:
        - name: Accept-Language
          in: header
          required: false
          schema:
            type: string
          example: uk, en;q=0.9
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchAnalyzeRequest'
      responses:
        '200':
          description: Batch analysis results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchAnalyzeResponse'
        '400':
          description: Validation error or batch too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/webhooks:
    get:
      summary: List webhook subscriptions
      tags: [Webhooks]
      description: Returns subscriptions without secrets.
      responses:
        '200':
          description: Subscription list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookListResponse'
    post:
      summary: Create webhook subscription
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
      responses:
        '201':
          description: Created subscription (includes secret once)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSubscription'
        '400':
          description: Validation error or subscription limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/webhooks/{id}:
    patch:
      summary: Enable or disable a webhook subscription
      tags: [Webhooks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWebhookRequest'
      responses:
        '200':
          description: Updated
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete webhook subscription
      tags: [Webhooks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Deleted
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/history:
    get:
      summary: Recent analysis reports (metadata only)
      tags: [Analysis]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated history list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'

  /api/v1/metrics:
    get:
      summary: Runtime metrics snapshot
      tags: [System]
      responses:
        '200':
          description: Metrics (p50/p95, cache hit, AI usage, errors)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetricsSnapshot'

components:
  schemas:
    AnalyzeRequest:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          minLength: 1
          maxLength: 32000
          description: Text to analyze
        mode:
          type: string
          enum: [quick, detailed]
          default: quick
        language:
          type: string
          enum: [en, uk]
          description: Report language override

    BatchAnalyzeItem:
      type: object
      required: [prompt]
      properties:
        id:
          type: string
          maxLength: 128
          description: Optional client correlation id
        prompt:
          type: string
          minLength: 1
          maxLength: 32000
        mode:
          type: string
          enum: [quick, detailed]
        language:
          type: string
          enum: [en, uk]

    BatchAnalyzeRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/BatchAnalyzeItem'
        mode:
          type: string
          enum: [quick, detailed]
          default: quick
          description: Default mode for items without their own mode
        language:
          type: string
          enum: [en, uk]
          description: Default report language for items without their own language

    BatchAnalyzeResponse:
      type: object
      properties:
        batchId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        results:
          type: array
          items:
            oneOf:
              - type: object
                required: [index, ok, result]
                properties:
                  index:
                    type: integer
                  id:
                    type: string
                  ok:
                    type: boolean
                    enum: [true]
                  result:
                    $ref: '#/components/schemas/AnalysisResult'
              - type: object
                required: [index, ok, error]
                properties:
                  index:
                    type: integer
                  id:
                    type: string
                  ok:
                    type: boolean
                    enum: [false]
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
        summary:
          type: object
          properties:
            total:
              type: integer
            succeeded:
              type: integer
            failed:
              type: integer
            processingTimeMs:
              type: integer

    LocalizedText:
      type: object
      properties:
        en:
          type: string
        uk:
          type: string

    AnalysisResult:
      type: object
      properties:
        analysisId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        rulesVersion:
          type: string
        risk:
          type: string
          enum: [SAFE, SUSPICIOUS, MALICIOUS]
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        confidence:
          type: integer
          minimum: 0
          maximum: 100
        action:
          type: string
          enum: [ALLOW, REVIEW, BLOCK, SANITIZE]
        language:
          type: string
          enum: [en, uk]
        detectedLanguage:
          type: string
          enum: [en, uk, mixed]
        explanation:
          $ref: '#/components/schemas/LocalizedText'
        educationalNote:
          $ref: '#/components/schemas/LocalizedText'
        recommendation:
          $ref: '#/components/schemas/LocalizedText'
        safeAlternative:
          $ref: '#/components/schemas/LocalizedText'
        aiInvoked:
          type: boolean
        pipelineStage:
          type: string
          enum: [exact, regex, ai, cache]

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        timestamp:
          type: string
          format: date-time
        version:
          type: string
        uptime:
          type: integer

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string

    HistoryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AnalysisHistoryItem'
        total:
          type: integer

    AnalysisHistoryItem:
      type: object
      properties:
        id:
          type: string
        risk:
          type: string
          enum: [SAFE, SUSPICIOUS, MALICIOUS]
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        confidence:
          type: integer
        action:
          type: string
          enum: [ALLOW, REVIEW, BLOCK, SANITIZE]
        promptLength:
          type: integer
        aiInvoked:
          type: boolean
        pipelineStage:
          type: string
          enum: [exact, regex, ai, cache]
        createdAt:
          type: string
          format: date-time

    MetricsSnapshot:
      type: object
      properties:
        totalRequests:
          type: integer
        errorCount:
          type: integer
        errorRate:
          type: number
        aiUsageRate:
          type: number
        cacheHitRate:
          type: number
        latencyMs:
          type: object
          properties:
            p50:
              type: number
            p95:
              type: number
            p99:
              type: number
        uptimeSeconds:
          type: integer

    WebhookEvent:
      type: string
      enum: [analysis.completed, analysis.blocked]

    CreateWebhookRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          maxLength: 2048
        secret:
          type: string
          minLength: 16
          maxLength: 256
          description: HMAC secret; auto-generated if omitted
        events:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEvent'
          minItems: 1

    UpdateWebhookRequest:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean

    WebhookSubscription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEvent'
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        secret:
          type: string
          description: Returned only on create

    WebhookListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WebhookSubscription'
        total:
          type: integer

    WebhookPayload:
      type: object
      description: |
        POST body sent to subscriber URLs. Signed with HMAC-SHA256 in
        `X-SHIELD-Signature: sha256=<hex>` using `timestamp.body`.
      properties:
        event:
          $ref: '#/components/schemas/WebhookEvent'
        timestamp:
          type: string
          format: date-time
        analysisId:
          type: string
          format: uuid
        promptHash:
          type: string
        promptLength:
          type: integer
        result:
          type: object
          properties:
            risk:
              type: string
              enum: [SAFE, SUSPICIOUS, MALICIOUS]
            severity:
              type: string
              enum: [LOW, MEDIUM, HIGH]
            confidence:
              type: integer
            action:
              type: string
              enum: [ALLOW, REVIEW, BLOCK, SANITIZE]
            categories:
              type: array
              items:
                type: string
            aiInvoked:
              type: boolean
            pipelineStage:
              type: string
              enum: [exact, regex, ai, cache]